pyproject.toml
[project]
name = "dbt-doctor"
version = "0.1.0"
description = "AI-driven quality & governance MCP Server for dbt projects. Audit coverage, profile data, detect schema drift, and auto-generate documentation."
authors = [{ name = "Orb" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["dbt", "mcp", "data-engineering", "llm", "ai", "schema", "documentation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "mcp[cli]>=1.3.0",
    "ruamel.yaml>=0.18.6",
    "psycopg[binary]>=3.1.0",
    "duckdb>=0.10.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
dbt-doctor = "dbt_doctor.server:main"

[project.urls]
Homepage = "https://github.com/yourusername/dbt-doctor"
Repository = "https://github.com/yourusername/dbt-doctor"
Issues = "https://github.com/yourusername/dbt-doctor/issues"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
